Move Totals() into metrics.go to fix the build on {Open,Net}BSD
