+++ internal/executor/metrics/metrics_test.go
@@ -6,6 +6,9 @@ import (
 	"context"
 	"fmt"
 	"github.com/cirruslabs/cirrus-ci-agent/internal/executor/metrics"
+	"github.com/shirou/gopsutil/cpu"
+	"github.com/shirou/gopsutil/mem"
+	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
 	"testing"
 	"time"
@@ -25,4 +28,18 @@ func TestMetrics(t *testing.T) {
 	require.Empty(t, result.Errors(), "we should never get errors here, but got %d", len(result.Errors()))
 	require.Len(t, result.ResourceUtilization.CpuChart, 4)
 	require.Len(t, result.ResourceUtilization.MemoryChart, 4)
+}
+
+func TestTotals(t *testing.T) {
+	ctx := context.Background()
+
+	expectedNumCpusTotal, err := cpu.Counts(true)
+	require.NoError(t, err)
+	expectedAmountMemory, err := mem.VirtualMemory()
+	require.NoError(t, err)
+
+	numCpusTotal, amountMemoryTotal, err := metrics.Totals(ctx)
+	require.NoError(t, err)
+	assert.EqualValues(t, expectedNumCpusTotal, numCpusTotal)
+	assert.EqualValues(t, expectedAmountMemory.Total, amountMemoryTotal)
 }
